+++ src/main/java/com/jeffpeng/jmod/nei/Plugin.java
@@ -1,10 +1,15 @@
 package com.jeffpeng.jmod.nei;
 
+import net.minecraft.item.ItemStack;
+import codechicken.nei.api.API;
+
 import com.jeffpeng.jmod.JMODPlugin;
 import com.jeffpeng.jmod.JMODPluginContainer;
+import com.jeffpeng.jmod.forgeevents.JMODHideItemStackEvent;
 
 import cpw.mods.fml.common.Loader;
 import cpw.mods.fml.common.event.FMLInitializationEvent;
+import cpw.mods.fml.common.eventhandler.SubscribeEvent;
 
 public class Plugin extends JMODPlugin {
 
@@ -14,7 +19,12 @@ public class Plugin extends JMODPlugin {
 	}
 	
 	public void on(FMLInitializationEvent event) {
-		if(Loader.isModLoaded("NotEnoughItem"))	new NEI_JMODConfig();
+		if(Loader.isModLoaded("NotEnoughItems"))	new NEI_JMODConfig();
+	}
+	
+	@SubscribeEvent
+	public void hideItemStack(JMODHideItemStackEvent event){
+		if(Loader.isModLoaded("NotEnoughItems"))	API.hideItem((ItemStack) event.get());
 	}
 
 }