+++ CHANGELOG.md
@@ -2,3 +2,4 @@ Changelog
 
 alpha1.3.1
 - First Version of the Plugin
+- Plugin listens now for JMODHideItemStackEvent to hide itemstacks on NEI